% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_tools.R
\name{carrega_dataset_ano}
\alias{carrega_dataset_ano}
\title{Carrega datasets por ano}
\usage{
carrega_dataset_ano(type, year)
}
\arguments{
\item{type}{O tipo do dataset. Valores possiveis: "raw", "wrangled", "spatial", "site"}

\item{year}{O ano dos datasets}
}
\description{
Função que captura os datasets por mes e mescla em um único dataset do ano.
}
\examples{
# Recupera o dataset de todos os crimes do ano de 2012
carrega_dataset_ano("spatial", 2012)
}
