% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_tools.R
\name{get_months}
\alias{get_months}
\title{Get Months}
\usage{
get_months(start = 1, end = 12)
}
\arguments{
\item{start}{O mês inicial (opcional)}

\item{end}{O mês final (opcional)}
}
\description{
Função auxiliar para recuperar um vetor de caracteres de meses,
muito útil para utilizar com a função mescla_datasets()
}
