% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_tools.R
\name{numero_nas}
\alias{numero_nas}
\title{Numero NAs}
\usage{
numero_nas(dataset)
}
\arguments{
\item{dataset}{O dataset que deseja analisar}
}
\description{
Conta o número de NAs em cada coluna de um dataset
}
