% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_tools.R
\name{recupera_datasets}
\alias{recupera_datasets}
\title{Recupera Datasets
Recupera o dataset de um ano especifico.}
\usage{
recupera_datasets(type, crime, years)
}
\arguments{
\item{type}{O tipo do dataset}

\item{crime}{O tipo do crime}

\item{years}{O ano que se deseja recuperar o dataset}
}
\description{
Recupera Datasets
Recupera o dataset de um ano especifico.
}
