% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_tools.R
\name{mescla_datasets}
\alias{mescla_datasets}
\title{Mescla datasets}
\usage{
mescla_datasets(type, crime, year, months)
}
\arguments{
\item{type}{O tipo do dataset. Valores possiveis: "raw", "wrangled", "spatial", "site"}

\item{crime}{O crime que se deseja mesclar os dados}

\item{year}{O ano dos datasets}
}
\description{
Funcao que captura os datasets por mes e mescla em um unico dataset do ano.
}
